+++ src/main/java/org/wildfly/security/sasl/entity/EntitySaslServer.java
@@ -19,7 +19,7 @@
 package org.wildfly.security.sasl.entity;
 
 import static org.wildfly.security._private.ElytronMessages.saslEntity;
-import static org.wildfly.security.asn1.ASN1.CONTEXT_SPECIFIC_MASK;
+import static org.wildfly.security.asn1.util.ASN1.CONTEXT_SPECIFIC_MASK;
 import static org.wildfly.security.sasl.entity.Entity.keyType;
 
 import java.security.InvalidKeyException;
