+++ src/test/java/org/wildfly/security/asn1/util/OidsUtilTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.wildfly.security.asn1;
+package org.wildfly.security.asn1.util;
 
 import org.junit.Assert;
 import org.junit.Test;
